requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
numpy==1.22.0
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.27.0
six==1.16.0
TA_Lib-0.4.23-cp38-cp38-win_amd64.whl
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
line_notify